sample-datapipeline-app/requirements.txt
azure-eventhub
python-dotenv>=0.5.1
azure-storage-blob